chore: remove replace home to app
