+++ src/App.css
@@ -11,7 +11,7 @@
 
 
 .App {
-  background-color: #282c34;
+  background-color: rgb(32, 35, 42);
   min-height: 100vh;
   display: flex;
   flex-direction: column;
@@ -32,4 +32,27 @@ body, main {
 
 .accordion {
   margin-bottom: 10px;
+}
+
+.form-control {
+  color: rgb(181, 175, 166) !important;
+  background-color: rgb(24, 26, 27) !important;
+  border-color: rgb(60, 65, 68) !important;
+}
+
+h2 {
+  background-color: rgb(0, 60, 60);
+  color: rgb(232, 230, 227);
+  width: 100vw;
+  margin-bottom: 0px !important;
+  border-radius: 6px;
+}
+
+.bg-dark {
+  background-color: rgb(44, 47, 49) !important
+}
+
+.dropdown-menu.show {
+  border: none;
+  background-color: #343A40 !important
 }
