+++ src/App.js
@@ -2,16 +2,94 @@ import React from "react";
 
 import "./App.css";
 import Header from "./Components/Navbar/Navbar";
-import Home from "./Components/Home/Home";
+import LatestInfo from "./Components/LatestInfo/LatestInfo";
+import Search from "./Components/Search/Search";
+import News from "./Components/News/News";
 import { withAuth0 } from "@auth0/auth0-react";
 
-function App() {
-  return (
-    <div className="App">
-      <Header />
-      <Home />
-    </div>
-  );
+const axios = require("axios");
+class App extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      haveSearched: false,
+      infoSearched: "",
+      coins: [],
+      searchTerms: [],
+    };
+  }
+
+  getCoins = async () => {
+    const response = await axios.get(
+      `${process.env.REACT_APP_CRYPTO_TRACKS_API}/coins`
+    );
+    const coins = response.data;
+    let searchTuples = coins.map((coin) => [coin.searchTerms, coin.symbol]);
+    let validSearchTerms = searchTuples.reduce((a, b) => a.concat(b[0]), []);
+    this.setState({
+      coins: searchTuples,
+      searchTerms: validSearchTerms,
+    });
+  };
+
+  getSymbol = (needle) => {
+    let match = this.state.coins.find((entry) => entry[0].includes(needle));
+    return match[1];
+  };
+
+  handleSearch = async (infoSearched) => {
+    if (
+      !infoSearched ||
+      !this.state.searchTerms.includes(infoSearched.toLowerCase())
+    ) {
+      console.warn("Invalid Search");
+      // TODO(stretch): Some Error Component
+    } else {
+      try {
+        // console.log(this.getSymbol(infoSearched));
+        let newResponse = await axios.get(
+          `${process.env.REACT_APP_CRYPTO_TRACKS_API}/news?q=${infoSearched}`
+        );
+        let coinResponse = await axios.get(
+          `${
+            process.env.REACT_APP_CRYPTO_TRACKS_API
+          }/coin-latest?symbol=${this.getSymbol(infoSearched)}`
+        );
+        // console.log("News Works: ", newResponse.data);
+        // console.log("Latest Price Works: ", coinResponse.data[0]);
+        this.setState({
+          newsResults: newResponse.data,
+          coinLatest: coinResponse.data[0],
+          haveSearched: true,
+        });
+      } catch (err) {
+        console.log(err);
+      }
+    }
+  };
+
+  async componentDidMount() {
+    await this.getCoins();
+  }
+
+  //TODO: Add Save / Delete Component Above Latest Info
+  render() {
+    return (
+      <div className="App">
+        <Header handleSearch={this.handleSearch} />
+        <Search
+          handleSearch={this.handleSearch}
+          suggestions={this.state.searchTerms}
+        />
+        {this.state.haveSearched ? (
+          <LatestInfo price={this.state.coinLatest} />
+        ) : (
+          ""
+        )}
+        {this.state.haveSearched ? <News news={this.state.newsResults} /> : ""}
+      </div>
+    );
+  }
 }
 
 export default withAuth0(App);
